Use random pivot index
So apparently when an array is already sorted the pivot index
matters. Choosing the first index each time in an already sorted
array creates a lot of extra work. Instead of trying to determine
the index choosing a random index is roughly a good (enough)
way of getting a good split on either side of the index.